Remove generated code for /cas1/placement-requests
