--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/api/cas1/PlacementRequestsCas1Delegate.kt
@@ -1,75 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.cas1
-
-import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
-import org.springframework.web.context.request.NativeWebRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1PlacementRequestDetail
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1PlacementRequestSummary
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1WithdrawPlacementRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.PlacementRequestRequestType
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.PlacementRequestSortField
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.PlacementRequestStatus
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.RiskTierLevel
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SortDirection
-import java.util.Optional
-
-@jakarta.annotation.Generated(value = ["org.openapitools.codegen.languages.KotlinSpringServerCodegen"], comments = "Generator version: 7.13.0")
-interface PlacementRequestsCas1Delegate {
-
-  fun getRequest(): Optional<NativeWebRequest> = Optional.empty()
-
-  fun getPlacementRequest(id: java.util.UUID): ResponseEntity<Cas1PlacementRequestDetail> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "{  \"desirableCriteria\" : [ null, null ],  \"booking\" : {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"characteristics\" : [ \"acceptsChildSexOffenders\", \"acceptsChildSexOffenders\" ],    \"premisesId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"premisesName\" : \"premisesName\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"departureDate\" : \"2000-01-23\",    \"type\" : \"space\",    \"arrivalDate\" : \"2000-01-23\"  },  \"notes\" : \"notes\",  \"assessmentDate\" : \"2000-01-23T04:56:07.000+00:00\",  \"assessor\" : {    \"telephoneNumber\" : \"telephoneNumber\",    \"roles\" : [ \"assessor\", \"assessor\" ],    \"probationDeliveryUnit\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"cruManagementArea\" : \"\",    \"cruManagementAreaDefault\" : \"\",    \"isActive\" : true,    \"version\" : 1,    \"cruManagementAreaOverride\" : \"\",    \"qualifications\" : [ \"pipe\", \"pipe\" ],    \"apArea\" : {      \"identifier\" : \"LON\",      \"name\" : \"Yorkshire & The Humber\",      \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"    },    \"service\" : \"service\",    \"permissions\" : [ \"cas1_adhoc_booking_create\", \"cas1_adhoc_booking_create\" ],    \"deliusUsername\" : \"deliusUsername\",    \"name\" : \"name\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"region\" : {      \"name\" : \"NPS North East Central Referrals\",      \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\"    },    \"email\" : \"email\"  },  \"openChangeRequests\" : [ {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"actualArrivalDate\" : \"2000-01-23\",    \"placementRequestId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"type\" : \"placementAppeal\"  }, {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"actualArrivalDate\" : \"2000-01-23\",    \"placementRequestId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"type\" : \"placementAppeal\"  } ],  \"type\" : \"normal\",  \"duration\" : 6,  \"essentialCriteria\" : [ \"isPIPE\", \"isPIPE\" ],  \"assessmentDecision\" : \"accepted\",  \"releaseType\" : \"licence\",  \"withdrawalReason\" : \"DuplicatePlacementRequest\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"radius\" : 0,  \"assessmentId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"spaceBookings\" : [ {    \"actualArrivalDate\" : \"2000-01-23\",    \"characteristics\" : [ null, null ],    \"isCancelled\" : true,    \"canonicalArrivalDate\" : \"2000-01-23\",    \"canonicalDepartureDate\" : \"2000-01-23\",    \"openChangeRequestTypes\" : [ \"placementAppeal\", \"placementAppeal\" ],    \"deliusEventNumber\" : \"deliusEventNumber\",    \"isNonArrival\" : true,    \"expectedDepartureDate\" : \"2000-01-23\",    \"keyWorkerAllocation\" : {      \"keyWorker\" : {        \"code\" : \"code\",        \"keyWorker\" : true,        \"name\" : \"Brown, James (PS - PSO)\"      },      \"allocatedAt\" : \"2000-01-23\"    },    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"premises\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"appealRequested\" : true,    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"actualDepartureDate\" : \"2000-01-23\",    \"plannedTransferRequested\" : true,    \"status\" : \"arrivingWithin6Weeks\"  }, {    \"actualArrivalDate\" : \"2000-01-23\",    \"characteristics\" : [ null, null ],    \"isCancelled\" : true,    \"canonicalArrivalDate\" : \"2000-01-23\",    \"canonicalDepartureDate\" : \"2000-01-23\",    \"openChangeRequestTypes\" : [ \"placementAppeal\", \"placementAppeal\" ],    \"deliusEventNumber\" : \"deliusEventNumber\",    \"isNonArrival\" : true,    \"expectedDepartureDate\" : \"2000-01-23\",    \"keyWorkerAllocation\" : {      \"keyWorker\" : {        \"code\" : \"code\",        \"keyWorker\" : true,        \"name\" : \"Brown, James (PS - PSO)\"      },      \"allocatedAt\" : \"2000-01-23\"    },    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"premises\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"appealRequested\" : true,    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"actualDepartureDate\" : \"2000-01-23\",    \"plannedTransferRequested\" : true,    \"status\" : \"arrivingWithin6Weeks\"  } ],  \"requestType\" : \"parole\",  \"isParole\" : true,  \"expectedArrival\" : \"2000-01-23\",  \"risks\" : {    \"mappa\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"level\" : \"level\"      }    },    \"tier\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"level\" : \"level\"      }    },    \"roshRisks\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"overallRisk\" : \"overallRisk\",        \"riskToChildren\" : \"riskToChildren\",        \"riskToPublic\" : \"riskToPublic\",        \"riskToKnownAdult\" : \"riskToKnownAdult\",        \"riskToStaff\" : \"riskToStaff\"      },      \"status\" : \"retrieved\"    },    \"flags\" : {      \"value\" : [ \"value\", \"value\" ]    },    \"crn\" : \"crn\"  },  \"isWithdrawn\" : true,  \"application\" : {    \"assessmentDecisionDate\" : \"2000-01-23\",    \"data\" : \"{}\",    \"document\" : \"{}\",    \"licenceExpiryDate\" : \"2000-01-23\",    \"isPipeApplication\" : true,    \"isEsapApplication\" : true,    \"cruManagementArea\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"apArea\" : {      \"identifier\" : \"LON\",      \"name\" : \"Yorkshire & The Humber\",      \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"submittedAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"assessmentId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"createdByUserId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"personStatusOnSubmission\" : \"InCustody\",    \"isEmergencyApplication\" : true,    \"isWomensApplication\" : true,    \"caseManagerUserDetails\" : {      \"telephoneNumber\" : \"telephoneNumber\",      \"name\" : \"name\",      \"email\" : \"email\"    },    \"arrivalDate\" : \"2000-01-23T04:56:07.000+00:00\",    \"risks\" : {      \"mappa\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"level\" : \"level\"        }      },      \"tier\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"level\" : \"level\"        }      },      \"roshRisks\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"overallRisk\" : \"overallRisk\",          \"riskToChildren\" : \"riskToChildren\",          \"riskToPublic\" : \"riskToPublic\",          \"riskToKnownAdult\" : \"riskToKnownAdult\",          \"riskToStaff\" : \"riskToStaff\"        },        \"status\" : \"retrieved\"      },      \"flags\" : {        \"value\" : [ \"value\", \"value\" ]      },      \"crn\" : \"crn\"    },    \"caseManagerIsNotApplicant\" : true,    \"applicantUserDetails\" : {      \"telephoneNumber\" : \"telephoneNumber\",      \"name\" : \"name\",      \"email\" : \"email\"    },    \"person\" : {      \"type\" : \"FullPerson\",      \"crn\" : \"crn\"    },    \"status\" : \"started\"  },  \"person\" : {    \"type\" : \"FullPerson\",    \"crn\" : \"crn\"  },  \"location\" : \"B74\",  \"applicationId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"legacyBooking\" : {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"characteristics\" : [ \"acceptsChildSexOffenders\", \"acceptsChildSexOffenders\" ],    \"premisesId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"premisesName\" : \"premisesName\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"departureDate\" : \"2000-01-23\",    \"type\" : \"space\",    \"arrivalDate\" : \"2000-01-23\"  },  \"status\" : \"notMatched\",  \"applicationDate\" : \"2000-01-23T04:56:07.000+00:00\"}")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun search(
-    status: PlacementRequestStatus?,
-    crnOrName: String?,
-    tier: RiskTierLevel?,
-    arrivalDateStart: java.time.LocalDate?,
-    arrivalDateEnd: java.time.LocalDate?,
-    requestType: PlacementRequestRequestType?,
-    cruManagementAreaId: java.util.UUID?,
-    page: Int?,
-    sortBy: PlacementRequestSortField?,
-    sortDirection: SortDirection?,
-  ): ResponseEntity<List<Cas1PlacementRequestSummary>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"requestedPlacementDuration\" : 0,  \"personTier\" : \"personTier\",  \"firstBookingPremisesName\" : \"firstBookingPremisesName\",  \"firstBookingArrivalDate\" : \"2000-01-23\",  \"person\" : {    \"type\" : \"FullPerson\",    \"crn\" : \"crn\"  },  \"requestedPlacementArrivalDate\" : \"2000-01-23\",  \"applicationSubmittedDate\" : \"2000-01-23\",  \"isParole\" : true,  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"applicationId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"placementRequestStatus\" : \"matched\"}, {  \"requestedPlacementDuration\" : 0,  \"personTier\" : \"personTier\",  \"firstBookingPremisesName\" : \"firstBookingPremisesName\",  \"firstBookingArrivalDate\" : \"2000-01-23\",  \"person\" : {    \"type\" : \"FullPerson\",    \"crn\" : \"crn\"  },  \"requestedPlacementArrivalDate\" : \"2000-01-23\",  \"applicationSubmittedDate\" : \"2000-01-23\",  \"isParole\" : true,  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"applicationId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"placementRequestStatus\" : \"matched\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun withdrawPlacementRequest(
-    id: java.util.UUID,
-    body: Cas1WithdrawPlacementRequest?,
-  ): ResponseEntity<Cas1PlacementRequestDetail> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "{  \"desirableCriteria\" : [ null, null ],  \"booking\" : {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"characteristics\" : [ \"acceptsChildSexOffenders\", \"acceptsChildSexOffenders\" ],    \"premisesId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"premisesName\" : \"premisesName\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"departureDate\" : \"2000-01-23\",    \"type\" : \"space\",    \"arrivalDate\" : \"2000-01-23\"  },  \"notes\" : \"notes\",  \"assessmentDate\" : \"2000-01-23T04:56:07.000+00:00\",  \"assessor\" : {    \"telephoneNumber\" : \"telephoneNumber\",    \"roles\" : [ \"assessor\", \"assessor\" ],    \"probationDeliveryUnit\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"cruManagementArea\" : \"\",    \"cruManagementAreaDefault\" : \"\",    \"isActive\" : true,    \"version\" : 1,    \"cruManagementAreaOverride\" : \"\",    \"qualifications\" : [ \"pipe\", \"pipe\" ],    \"apArea\" : {      \"identifier\" : \"LON\",      \"name\" : \"Yorkshire & The Humber\",      \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"    },    \"service\" : \"service\",    \"permissions\" : [ \"cas1_adhoc_booking_create\", \"cas1_adhoc_booking_create\" ],    \"deliusUsername\" : \"deliusUsername\",    \"name\" : \"name\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"region\" : {      \"name\" : \"NPS North East Central Referrals\",      \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\"    },    \"email\" : \"email\"  },  \"openChangeRequests\" : [ {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"actualArrivalDate\" : \"2000-01-23\",    \"placementRequestId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"type\" : \"placementAppeal\"  }, {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"actualArrivalDate\" : \"2000-01-23\",    \"placementRequestId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"type\" : \"placementAppeal\"  } ],  \"type\" : \"normal\",  \"duration\" : 6,  \"essentialCriteria\" : [ \"isPIPE\", \"isPIPE\" ],  \"assessmentDecision\" : \"accepted\",  \"releaseType\" : \"licence\",  \"withdrawalReason\" : \"DuplicatePlacementRequest\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"radius\" : 0,  \"assessmentId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"spaceBookings\" : [ {    \"actualArrivalDate\" : \"2000-01-23\",    \"characteristics\" : [ null, null ],    \"isCancelled\" : true,    \"canonicalArrivalDate\" : \"2000-01-23\",    \"canonicalDepartureDate\" : \"2000-01-23\",    \"openChangeRequestTypes\" : [ \"placementAppeal\", \"placementAppeal\" ],    \"deliusEventNumber\" : \"deliusEventNumber\",    \"isNonArrival\" : true,    \"expectedDepartureDate\" : \"2000-01-23\",    \"keyWorkerAllocation\" : {      \"keyWorker\" : {        \"code\" : \"code\",        \"keyWorker\" : true,        \"name\" : \"Brown, James (PS - PSO)\"      },      \"allocatedAt\" : \"2000-01-23\"    },    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"premises\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"appealRequested\" : true,    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"actualDepartureDate\" : \"2000-01-23\",    \"plannedTransferRequested\" : true,    \"status\" : \"arrivingWithin6Weeks\"  }, {    \"actualArrivalDate\" : \"2000-01-23\",    \"characteristics\" : [ null, null ],    \"isCancelled\" : true,    \"canonicalArrivalDate\" : \"2000-01-23\",    \"canonicalDepartureDate\" : \"2000-01-23\",    \"openChangeRequestTypes\" : [ \"placementAppeal\", \"placementAppeal\" ],    \"deliusEventNumber\" : \"deliusEventNumber\",    \"isNonArrival\" : true,    \"expectedDepartureDate\" : \"2000-01-23\",    \"keyWorkerAllocation\" : {      \"keyWorker\" : {        \"code\" : \"code\",        \"keyWorker\" : true,        \"name\" : \"Brown, James (PS - PSO)\"      },      \"allocatedAt\" : \"2000-01-23\"    },    \"tier\" : \"tier\",    \"expectedArrivalDate\" : \"2000-01-23\",    \"person\" : {      \"personType\" : \"FullPersonSummary\",      \"crn\" : \"crn\"    },    \"premises\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"appealRequested\" : true,    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"actualDepartureDate\" : \"2000-01-23\",    \"plannedTransferRequested\" : true,    \"status\" : \"arrivingWithin6Weeks\"  } ],  \"requestType\" : \"parole\",  \"isParole\" : true,  \"expectedArrival\" : \"2000-01-23\",  \"risks\" : {    \"mappa\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"level\" : \"level\"      }    },    \"tier\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"level\" : \"level\"      }    },    \"roshRisks\" : {      \"value\" : {        \"lastUpdated\" : \"2000-01-23\",        \"overallRisk\" : \"overallRisk\",        \"riskToChildren\" : \"riskToChildren\",        \"riskToPublic\" : \"riskToPublic\",        \"riskToKnownAdult\" : \"riskToKnownAdult\",        \"riskToStaff\" : \"riskToStaff\"      },      \"status\" : \"retrieved\"    },    \"flags\" : {      \"value\" : [ \"value\", \"value\" ]    },    \"crn\" : \"crn\"  },  \"isWithdrawn\" : true,  \"application\" : {    \"assessmentDecisionDate\" : \"2000-01-23\",    \"data\" : \"{}\",    \"document\" : \"{}\",    \"licenceExpiryDate\" : \"2000-01-23\",    \"isPipeApplication\" : true,    \"isEsapApplication\" : true,    \"cruManagementArea\" : {      \"name\" : \"name\",      \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"    },    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"apArea\" : {      \"identifier\" : \"LON\",      \"name\" : \"Yorkshire & The Humber\",      \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"    },    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"submittedAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"assessmentId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"createdByUserId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"personStatusOnSubmission\" : \"InCustody\",    \"isEmergencyApplication\" : true,    \"isWomensApplication\" : true,    \"caseManagerUserDetails\" : {      \"telephoneNumber\" : \"telephoneNumber\",      \"name\" : \"name\",      \"email\" : \"email\"    },    \"arrivalDate\" : \"2000-01-23T04:56:07.000+00:00\",    \"risks\" : {      \"mappa\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"level\" : \"level\"        }      },      \"tier\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"level\" : \"level\"        }      },      \"roshRisks\" : {        \"value\" : {          \"lastUpdated\" : \"2000-01-23\",          \"overallRisk\" : \"overallRisk\",          \"riskToChildren\" : \"riskToChildren\",          \"riskToPublic\" : \"riskToPublic\",          \"riskToKnownAdult\" : \"riskToKnownAdult\",          \"riskToStaff\" : \"riskToStaff\"        },        \"status\" : \"retrieved\"      },      \"flags\" : {        \"value\" : [ \"value\", \"value\" ]      },      \"crn\" : \"crn\"    },    \"caseManagerIsNotApplicant\" : true,    \"applicantUserDetails\" : {      \"telephoneNumber\" : \"telephoneNumber\",      \"name\" : \"name\",      \"email\" : \"email\"    },    \"person\" : {      \"type\" : \"FullPerson\",      \"crn\" : \"crn\"    },    \"status\" : \"started\"  },  \"person\" : {    \"type\" : \"FullPerson\",    \"crn\" : \"crn\"  },  \"location\" : \"B74\",  \"applicationId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"legacyBooking\" : {    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",    \"characteristics\" : [ \"acceptsChildSexOffenders\", \"acceptsChildSexOffenders\" ],    \"premisesId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"premisesName\" : \"premisesName\",    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",    \"departureDate\" : \"2000-01-23\",    \"type\" : \"space\",    \"arrivalDate\" : \"2000-01-23\"  },  \"status\" : \"notMatched\",  \"applicationDate\" : \"2000-01-23T04:56:07.000+00:00\"}")
-          break
-        }
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "{  \"instance\" : \"f7493e12-546d-42c3-b838-06c12671ab5b\",  \"detail\" : \"You provided invalid request parameters\",  \"type\" : \"https://example.net/validation-error\",  \"title\" : \"Invalid request parameters\",  \"status\" : 400}")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-}
